+++ libs/core/src/components/context-menu/context-menu.ts
@@ -1,4 +1,4 @@
-import { HTMLTemplateResult, LitElement, nothing } from 'lit'
+import { HTMLTemplateResult, nothing } from 'lit'
 import { msg } from '@lit/localize'
 import { classMap } from 'lit-html/directives/class-map.js'
 import { property, queryAsync, state } from 'lit/decorators.js'
@@ -14,7 +14,8 @@ import { TransitionalStyles } from '../../transitional-styles'
 
 import styles from './context-menu.styles'
 
-import '../../primitives/menu/menu'
+import '../../primitives/menu'
+import '../../primitives/popover'
 
 /**
  * @element gds-context-menu