+++ libs/core/src/components/datepicker/date-part-spinner.ts
@@ -1,11 +1,7 @@
 import { LitElement } from 'lit'
 import { property, state } from 'lit/decorators.js'
-import {
-  gdsCustomElement,
-  html,
-} from '../../utils/helpers/custom-element-scoping'
-import { watch } from '../../utils/decorators'
-import { Focusable } from '../../mixins/focusable'
+import { gdsCustomElement, html } from '../../scoping'
+import { watch } from '../../utils/decorators/watch'
 
 /**
  * @element gds-date-part-spinner