+++ libs/core/src/components/grid/grid.ts
@@ -1,5 +1,4 @@
-import { css, CSSResult, LitElement, unsafeCSS } from 'lit'
-import type { CSSResultGroup } from 'lit'
+import { css, LitElement, unsafeCSS } from 'lit'
 
 import { property, state } from 'lit/decorators.js'
 import {
@@ -7,11 +6,10 @@ import {
   html,
 } from '../../utils/helpers/custom-element-scoping'
 
-import { watch } from '../../utils/decorators'
+import { watch } from '../../utils/decorators/watch'
 
 import { tokens } from '../../tokens.style'
 import GridCSS from './grid.style.css'
-// const gridCSS = unsafeCSS(grid)
 
 type GridSizes = 'none' | 'xs' | 's' | 'm' | 'l' | 'xl' | '2xl' | '3xl'
 const BreakpointPattern =