+++ libs/core/src/primitives/menu/menu-heading.ts
@@ -1,6 +1,6 @@
 import { html } from 'lit'
 import { GdsElement } from '../../gds-element'
-import { gdsCustomElement } from '../../utils/helpers/custom-element-scoping'
+import { gdsCustomElement } from '../../scoping'
 import { TransitionalStyles } from '../../transitional-styles'
 
 import style from './menu-heading.styles'