+++ libs/core/web-test-runner.config.mjs
@@ -1,0 +1,11 @@
+export default {
+  files: ['dist/tests/core/**/*.test.js'],
+  nodeResolve: true,
+  concurrentBrowsers: 3,
+  testFramework: {
+    config: {
+      timeout: 3000,
+      retries: 1,
+    },
+  },
+}