+++ libs/react/src/lib/grouped-list/grouped-list.ts
@@ -1,6 +1,7 @@
 import React from 'react'
 import { createComponent } from '@lit/react'
-import { GdsGroupedList, getScopedTagName } from '@sebgroup/green-core'
+import { GdsGroupedList } from '@sebgroup/green-core/components/grouped-list/index.js'
+import { getScopedTagName } from '@sebgroup/green-core/scoping'
 import { registerTransitionalStyles } from '@sebgroup/green-core/transitional-styles'
 
 registerTransitionalStyles()