+++ libs/tokens/src/transforms.ts
@@ -77,6 +77,18 @@ const transforms: Record<
       return value.setAlpha(token.alpha).toString()
     },
   },
+  'green/color/UIColorSwift': {
+    name: 'green/color/UIColorSwift',
+    type: 'value',
+    matcher: (token) => token.path.includes('color'),
+    transformer: function (token) {
+      const { r, g, b, a } = tinycolor(token.value).toRgb()
+      const rFixed = (r / 255.0).toFixed(3)
+      const gFixed = (g / 255.0).toFixed(3)
+      const bFixed = (b / 255.0).toFixed(3)
+      return `UIColor(red: ${rFixed}, green: ${gFixed}, blue: ${bFixed}, alpha: ${a})`
+    },
+  },
 }
 
 export default transforms